refactor: reorg rest-api folders
