+++ apps/rest/src/app.module.ts
@@ -1,12 +1,15 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { AppService } from './services/app';
 import CustomTypeOrmModule from './providers/typeorm.module';
 import entities from './models';
+import { AppController } from './controllers/app';
 @Module({
   imports: [
-    ConfigModule.forRoot({ isGlobal: true }),
+    ConfigModule.forRoot({
+      isGlobal: true,
+      envFilePath: process.env.NODE_ENV === 'test' ? '.env.test' : '.env',
+    }),
     CustomTypeOrmModule.forRootAsync(null, {
       entities,
       // FIXME: using db migration instead of synchronize
