+++ apps/rest/src/controllers/app.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { AppController } from './app';
+import { AppService } from '../services/app';
 
 describe('AppController', () => {
   let appController: AppController;
