+++ apps/rest/src/controllers/app.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { AppService } from './app.service';
+import { AppService } from '../services/app';
 
 @Controller()
 export class AppController {
